Faas 7: tõsta save metadata git ja bulk routerisse #36
